requirements.txt
# HTTP Client
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Web Automation
selenium==4.15.0
webdriver-manager==4.0.1

# HTML Parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# Retry Logic
tenacity==8.2.3

# Logging
colorlog==6.7.0

# Utilities
python-dateutil==2.8.2
